Rename game_grid files
